refactor: lib/interfaces/ディレクトリを削除してtypes/api.tsとfeatures/receipt/types.tsに移動
